refactor: 기보 실행 테스트 패키지 위치 변경
